Move specs for school_administrator_dashboard, move HighchartsWrapper and Datepicker into components
