+++ app/assets/javascripts/components/HighchartsWrapper.js
