+++ app/assets/javascripts/school_administrator_dashboard/DashResetButton.js
