+++ app/assets/javascripts/school_administrator_dashboard/DashboardHelpers.js
