+++ app/assets/javascripts/school_administrator_dashboard/DashboardLoader.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
+import GenericLoader from '../components/GenericLoader';
+import {apiFetchJson} from '../helpers/apiFetchJson';
 import SchoolwideAbsences from './absences_dashboard/SchoolwideAbsences';
 import SchoolwideTardies from './tardies_dashboard/SchoolwideTardies';
 import SchoolwideDisciplineIncidents from './discipline_dashboard/SchoolwideDisciplineIncidents';
-import GenericLoader from '../../components/GenericLoader';
-import {apiFetchJson} from '../../helpers/apiFetchJson';
+
 
 class DashboardLoader extends React.Component {
   constructor(props) {
