+++ app/assets/javascripts/school_administrator_dashboard/DateSlider.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Slider from 'rc-slider';
 import moment from 'moment';
-import Datepicker from '../../student_profile/Datepicker';
+import Datepicker from '../components/Datepicker';
 
 const createSliderWithTooltip = Slider.createSliderWithTooltip;
 const Range = createSliderWithTooltip(Slider.Range);
