+++ app/assets/javascripts/school_administrator_dashboard/DateSlider.test.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import moment from 'moment';
 import { shallow } from 'enzyme';
-
-import DateSlider from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/DateSlider';
+import DateSlider from './DateSlider';
 
 describe('Dashboard Date Slider', () => {
   const slider = shallow(<DateSlider
