+++ app/assets/javascripts/school_administrator_dashboard/absences_dashboard/SchoolAbsenceDashboard.test.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import moment from 'moment';
 import { shallow } from 'enzyme';
-
-import SchoolAbsenceDashboard from '../../../app/assets/javascripts/school_administrator_dashboard/dashboard_components/absences_dashboard/SchoolAbsenceDashboard';
 import {
   createTestEvents,
   createStudents
-} from './DashboardTestData';
+} from '../DashboardTestData';
+import SchoolAbsenceDashboard from './SchoolAbsenceDashboard';
 
 
 describe('SchoolAbsenceDashboard', () => {
