+++ app/assets/javascripts/school_administrator_dashboard/tardies_dashboard/SchoolTardiesDashboard.js
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 import moment from 'moment';
+import {latestNoteDateText} from '../../helpers/latestNoteDateText';
+import DesignChangesBanner from '../../components/DesignChangesBanner';
 import DashboardHelpers from '../DashboardHelpers';
 import StudentsTable from '../StudentsTable';
 import DashboardBarChart from '../DashboardBarChart';
-import {latestNoteDateText} from '../../../helpers/latestNoteDateText';
 import DashRangeButtons from '../DashRangeButtons';
-import DesignChangesBanner from '../../../components/DesignChangesBanner';
 
 class SchoolTardiesDashboard extends React.Component {
