+++ app/assets/javascripts/school_administrator_dashboard/tardies_dashboard/SchoolwideTardies.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import moment from 'moment';
+import { shallow } from 'enzyme';
+import {createStudents} from '../DashboardTestData';
+import SchoolwideTardies from './SchoolwideTardies';
+
+
+it('renders the school tardies dashboard', () => {
+  const tardies = shallow(<SchoolwideTardies dashboardStudents={createStudents(moment.utc())} />);
+  expect(tardies.find('SchoolTardiesDashboard').length).toEqual(1);
+});
